// components/services-block/enums.ts
export enum CardType {
  PRIMARY,
  SECONDARY,
}

export enum PrimaryContent {
  HEADER = "Чек-листы и документы для аренды",
  DESC = "Договора защищающие интересы арендатора. Чек-листы по выбору объекта, осмотру и переговорам помогут не пропустить важное",
  BUTTON_TEXT = "Купить за 799 ₽",
}

export enum SecondaryContent {
  HEADER = "Пакет бесплатных документов",
  DESC = "Пригодятся от подписания договора аренды до его прекращения. Документы универсальны для всех участников сделки",
  BUTTON_TEXT = "Получить документы",
}